Investment is also able to go back.

+++ InvestmentArchive.py
@@ -16,7 +16,7 @@ class InvestmentArchive:
         self.SetUpConnection()
         self.c.execute('''
               CREATE TABLE IF NOT EXISTS ArchiveInvestment                                                         
-              ([Investment_ID] VARCHAR PRIMARY KEY,[User_ID] VARCHAR, [Gold] Real , [Purity] Real, [BoughtFor] REAL,[ProfitLoss] Real,[deleted_at] TIMESTAMP DEFAULT CURRENT_TIMESTAMP)
+              ([Investment_ID] VARCHAR ,[User_ID] VARCHAR, [Gold] Real , [Purity] Real, [BoughtFor] REAL,[ProfitLoss] Real,[deleted_at] TIMESTAMP DEFAULT CURRENT_TIMESTAMP)
               ''')
         self.conn.commit()
         self.conn.close()
@@ -44,26 +44,30 @@ class InvestmentArchive:
         finally:
             self.conn.close()
 
-    def getData(self, *Investment_ID):
+    def getData(self, User_ID, *Investment_ID):
         self.SetUpConnection()
-        self.c.execute("SELECT COUNT(*) FROM ArchiveInvestment")
+        Data = None
+        self.c.execute("SELECT COUNT(*) FROM ArchiveInvestment WHERE User_ID = ?", (User_ID,))
         Count = self.c.fetchone()[0]
         print(Count)
-
-        if not Investment_ID:
-            self.c.execute("SELECT * FROM ArchiveInvestment LIMIT 1 OFFSET ?", (Count - 1,))
-        else:
-            self.c.execute("SELECT * FROM ArchiveInvestment WHERE Investment_ID = ? LIMIT 1 OFFSET ?", (Investment_ID, Count - 1,))
-        Data = self.c.fetchone()
-        if not Investment_ID:
-            # needs id to delete, its fine as id is primary key and unique.
-            self.c.execute(
-                "DELETE FROM ArchiveInvestment WHERE Investment_ID IN (SELECT Investment_ID FROM ArchiveInvestment LIMIT 1 OFFSET ?)",
-                (Count - 1,))
-        else:
-            self.c.execute(
-                "DELETE FROM ArchiveUser WHERE (SELECT * FROM ArchiveInvestment WHERE User_ID = ? LIMIT 1 OFFSET ?)",
-                (Investment_ID, Count - 1,))
+        if Count > 0:
+            if not Investment_ID:
+                self.c.execute("SELECT * FROM ArchiveInvestment WHERE User_ID = ? LIMIT 1 OFFSET ?",
+                               (User_ID, Count - 1,))
+            else:
+                self.c.execute(
+                    "SELECT * FROM ArchiveInvestment WHERE Investment_ID = ? AND User_ID = ? LIMIT 1 OFFSET ?",
+                    (Investment_ID, User_ID, Count - 1,))
+            Data = self.c.fetchone()
+            if not Investment_ID:
+                # needs id to delete, its fine as id is primary key and unique.
+                self.c.execute(
+                    "DELETE FROM ArchiveInvestment WHERE Investment_ID IN (SELECT Investment_ID FROM ArchiveInvestment WHERE User_ID = ? LIMIT 1 OFFSET ?)",
+                    (User_ID, Count - 1,))
+            else:
+                self.c.execute(
+                    "DELETE FROM ArchiveUser WHERE (SELECT * FROM ArchiveInvestment WHERE User_ID = ? LIMIT 1 OFFSET ?)",
+                    (User_ID, Count - 1,))
 
         self.conn.commit()
         self.conn.close()
